requirements.txt
fastapi
uvicorn
openai==0.28.1
python-dotenv
pandas